Start adding support for PHI operator in Beanstalk
Summary:
BMG supports a unary operator PHI which computes the CDF of the standard normal distribution; it takes a real and returns the probability that a sample will be below the argument.

We will need to support compiling models written in Python into BMG graphs which use this operator; as usual, I'm starting by adding a node type and doing basic testing on it.

In upcoming diffs we will add it to the graph builder proper, and then write a detector which turns things like

x = Normal(0,1).cdf(some_sample)

into PHI nodes.

Reviewed By: wtaha

Differential Revision: D23797901

fbshipit-source-id: cc951a824b6720e71550977496d9052363bbca24