Start adding support for PHI operator in Beanstalk (#393)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/393

BMG supports a unary operator PHI which computes the CDF of the standard normal distribution; it takes a real and returns the probability that a sample will be below the argument.

We will need to support compiling models written in Python into BMG graphs which use this operator; as usual, I'm starting by adding a node type and doing basic testing on it.

In upcoming diffs we will add it to the graph builder proper, and then write a detector which turns things like

x = Normal(0,1).cdf(some_sample)

into PHI nodes.

Reviewed By: wtaha

Differential Revision: D23797901

fbshipit-source-id: 98391fa4b4e8b52af321f1e20ebe7f5ac5d89b9d