Finish adding support for PHI operator in Beanstalk
Summary:
We now detect usages of Normal(0.0, 1.0).cdf(x) in Python models and generate them as PHI nodes in BMG.

Note that in the test we show that a to-real node is inserted between the positive real sample and the PHI, which requires a real input. Also note that PHI outputs a probability, which is the input to the Bernoulli.

Reviewed By: wtaha

Differential Revision: D23800138

fbshipit-source-id: e36b8dd424e0ab57a8841e96427187b939ed46c7