Relaxed delta in tests that broke due to correct definition of norm (#397)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/397

In a previous diff, a typo in the definition of the L_1 norm was fixed (D23840917 (https://github.com/facebookincubator/beanmachine/commit/e53b02fbeccf98b14b42e01ec556dfad65e92be0)). This led to several existing tests breaking. This diff temporarily fixes those tests by relaxing the delta that they use for comparing the resulting average. The plan is to rewrite tests to use p-values, and to introduce this in upcoming diffs.

Reviewed By: nimar

Differential Revision: D23933801

fbshipit-source-id: 3d99d1142ae60d75fdee88a3e8542db401a9af8d