Fix uninitialized memory bug in BMG (#398)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/398

The constructors for graph nodes in BMG need to initialize both the type and the value of a graph upon construction; two of the nodes were initializing only the type.

The reason we must initialize the value is because the default zero value needs to be shown when the graph is created and then represented as a string; the values are output then. If we do not initialize the values then test cases which check the string form of the graph can fail unexpectedly.

Reviewed By: dengdetian0603

Differential Revision: D23966395

fbshipit-source-id: 3d3e17e051cb4742f326d3e588166c9479220acd