Automatically eliminate identity multiplication and addition
Summary:
If during graph accumulation we have an addition of zero to something, or a multiplication by one, or a multiplication by zero, we can optimize it away immediately. This saves having to deal with type analysis, later optimization passes, and so on.

The logistic regression model that I recently added as a test case has multiplication by 1.0 in it, and so its generated graph became smaller as a result.

The multiplication node unit test was using multiplication by 1.0 as a test case, so I updated parts of it to multiply by 2.0 instead, so that the multiplication would not be optimized away for the test.

Reviewed By: wtaha

Differential Revision: D23831426

fbshipit-source-id: 4547f20a1627ee2a2c1c0092b8d2be813071224f