Bug fix for missing ast.Expr around an ast.Call in a previous diff (#401)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/401

Eric kindly reported an issue with single-static assignment rewrite rules for comprehensions through a disabled diff (D23970806). This diff fixes the bug by adding this constructor. In addition, the types of the functions involved were refined to be more precise and to help avoid such problems in the future. The type refinements also helped in deciding on the best place to fix this problem. Finally, the old tests that this fix breaks are updated

Reviewed By: ericlippert

Differential Revision: D24031665

fbshipit-source-id: 0ccfc05f6e8aac030c48a7a2b271a9316769261a