Start implementing support for negative reals in BMG (#403)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/403

We are exploring the notion of adding a "negative real" type to the BMG type system. This is particularly useful because a negative real number is a log-probability; we can modify the type system so that exp(neg_real) is a probability, and log(prob) is a negative real. (We can also eliminate the neg_log operation which says that the negative of a log(prob) is a positive real.)

This stack of diffs is for prototyping up this change; in this first diff we just add the type to the type system and add the ability to create constant negative real nodes.

I will add support for this type to the various operators one at a time, and also add support to Beanstalk along the way.

Reviewed By: nimar

Differential Revision: D24062964

fbshipit-source-id: 620e83eaf4f701f7e8c361fe7a85cc03ded7a7f9