Continue adding negative real type to Beanstalk (#411)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/411

In the previous diff we added zero and negative real types to the beanstalk type lattice. In this diff we start classifying constant values as zero or negative real.

The idea here is that we need to know whether a given node can be used in a context where a particular type is required; the easiest nodes to analyze are constants so we start with them. A node that is the constant zero can be used in a context that requires a bool, but can also be used in a context that requires a negative real, so we need the zero type to represent "I could be either".  (This is analogous to how we are already using the "one" type to represent "I could be a bool or a simplex")

A nice property of this change is that previously, we would classify real constants as either "positive real" if positive, or "real" if negative, which seems a little odd. We now classify positives as positives and negatives as negatives.  Similarly, we would classify true as "one" and false as "bool"; we now classify true as "one" and false as "zero", which makes a little more sense.

Reviewed By: wtaha

Differential Revision: D24171008

fbshipit-source-id: abf723b428f7b509cac4237a2255daeb3bff0c1f