BMG LOG operator now supports probability operand
Summary:
The BMG LOG operator previously supported only taking an input of type positive real and producing an output of type real.

We now also support taking an input of type probability and producing an output of type negative real.

I will add support for this to the beanstalk compiler in the next diff.

Reviewed By: wtaha

Differential Revision: D24203997

fbshipit-source-id: e1478886e99bc44607f966afc51dae3084905e64