Beanstalk log operator now supports probability operands
Summary:
We are continuing with the process of adding support for negative reals to the BMG type system. In the previous diff we updated BMG log operator to support probability operands; in this diff we similarly add support to beanstalk.

That's going to cause churn in our tests which verify -log(prob) -> positive real so I've disabled that test; it will get turned on again in a few diffs.

Reviewed By: wtaha

Differential Revision: D24205043

fbshipit-source-id: 2cb65b60eedfd6bf1509ee42ea95dd58ac44d05d