Beanstalk log operator now supports probability operands (#418)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/418

We are continuing with the process of adding support for negative reals to the BMG type system. In the previous diff we updated BMG log operator to support probability operands; in this diff we similarly add support to beanstalk.

That's going to cause churn in our tests which verify -log(prob) -> positive real so I've disabled that test; it will get turned on again in a few diffs.

Reviewed By: wtaha

Differential Revision: D24205043

fbshipit-source-id: 8bac67f9c281f16b8a62e6b31a5ca8acd3a927a8