BMG NEGATE operator now supports negative and positive real operands
Summary:
This diff was originally supposed to add support for neg_real in the NEGATE operator, but Detian already did that in another diff.

I'm keeping this diff though because it adds a few test cases and tweaks the wording of some error messages to make them more consistent.

In an upcoming diff I will add negative real support for the negate operator to the beanstalk compiler's graph builder.

Reviewed By: wtaha

Differential Revision: D24206016

fbshipit-source-id: 9e8f8b9a067a2b9d3f84dcdae700775fcafc575a