BMG NEGATE operator now supports negative and positive real operands (#423)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/423

This diff was originally supposed to add support for neg_real in the NEGATE operator, but Detian already did that in another diff.

I'm keeping this diff though because it adds a few test cases and tweaks the wording of some error messages to make them more consistent.

In an upcoming diff I will add negative real support for the negate operator to the beanstalk compiler's graph builder.

Reviewed By: wtaha

Differential Revision: D24206016

fbshipit-source-id: 39bf3737ea9a47d907c357656f9e7a78804249b6