Beanstalk negate operator now supports negative, positive reals (#424)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/424

In the previous diff we made the BMG negation operator support real, negative real and positive real operands. In this diff we do the same to the beanstalk type analyzer. Note that this causes some minor test changes in the test which verifies that (1 + (-prob)) is correctly turned into a complement operator, since the type analysis of the -prob is now different.

We now have enough gear in place that we can remove the NEG_LOG operator from Beanstalk and BMG. That will be the next few diffs.

Reviewed By: wtaha

Differential Revision: D24207146

fbshipit-source-id: 92e061a31a4810df50fef7a31403406ed277c8f7