Start re-enabling disabled tests
Summary:
There are a bunch of disabled codegen tests; we should turn those back on.

In this diff I've started pulling the disabled test cases apart into individual unit tests and selectively re-enabling them.

Note that the `dot_3` output changed because we now optimize away the "add zero to a sum" operation; the zero node becomes an orphan.

I'll gradually re-enable the rest of these tests in future diffs.

Reviewed By: wtaha

Differential Revision: D24408403

fbshipit-source-id: 9f11c683e985b1e2e2516180eaebf10cbd13ea8e