Start re-enabling disabled tests (#427)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/427

There are a bunch of disabled codegen tests; we should turn those back on.

In this diff I've started pulling the disabled test cases apart into individual unit tests and selectively re-enabling them.

Note that the `dot_3` output changed because we now optimize away the "add zero to a sum" operation; the zero node becomes an orphan.

I'll gradually re-enable the rest of these tests in future diffs.

Reviewed By: wtaha

Differential Revision: D24408403

fbshipit-source-id: 367bca16a290e5b58f1f2e60a98ebb952413aa35