Avoids divid-by-zero in NUTS
Summary:
In current implementation of NUTS proposer, it's possible to raise zero to a negative power when  `acceptance_prob == 0.0` and  `a == -1`.

This undefine behavior would usually be evaluated to NaN when running in Buck's opt mode (or in regular conda/pip). However, in dev mode, Buck's `UndefinedBehaviorSanitizer` is going to complain about this usage and terminates any program that uses the proposer.

Differential Revision: D24495518

fbshipit-source-id: 0ff3af2e4cd093c3cc50abba453f951e1eccfe09