Avoids divid-by-zero in NUTS (#430)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/430

In current implementation of NUTS proposer, it's possible to raise zero to a negative power when  `acceptance_prob == 0.0` and  `a == -1`.

This undefine behavior would usually be evaluated to inf when running in Buck's opt mode (or in regular conda/pip). However, in dev mode, Buck's `UndefinedBehaviorSanitizer` is going to complain about this usage and terminates any program that uses the proposer.

Reviewed By: yucenli

Differential Revision: D24495518

fbshipit-source-id: df127aa92b1e7dc3da336205ebd9428e64db9d7c