Start removing negative log feature (#433)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/433

Now that we have a type analysis that shows that log of a probability is a negative real, we can remove the "negative log" node. It's purpose was only to allow us to represent log probabilities as positive reals.

I've started by re-enabling the test that showed how neg(log(p)) was transformed into neglog(p), and updating it to show that we do a correct type analysis.

I will remove the rest of the code for this from Beanstalk and BMG in the next few diffs.

Reviewed By: wtaha

Differential Revision: D24338918

fbshipit-source-id: 634513043eb7e412f51456d3ff5813421945763d