Start implementing comparison operators in compiler
Summary:
BMG does not support comparison operators -- yet.  But we need to track them in the graph accumulator for several reasons.

(1) We need to give errors! If someone tries to translate a model to BMG that uses comparison operators on graph nodes, we need to clearly and accurately describe the error.

(2) We might have model analysis scenarios other than translating a model to BMG which require correct analysis of all the operations in the model

(3) We might support comparison operators in BMG; when we do, it will be easy to modify the existing code to support those operations.

In this diff I simply add classes representing `<`, `<=`, `>`, `>=` and `==` to the set of supported nodes. In later diffs I will add code to accumulate these into the graph and produce errors.

Reviewed By: wtaha

Differential Revision: D24376044

fbshipit-source-id: 448ecea1fd8ea0ddd2d4b092df3a42166c7cb94a