Give compiler errors for stochastic comparisons
Summary: We need to give an error when an attempt is made to compare anything with a stochastic node, because we have no ability at this time to compile it into a BMG graph. Here we turn statements of the form  `x = y op z` for op being any of `< > <= >= == !=` into graph nodes in the graph accumulator, and then detect in the problem fixer whether any such nodes remain and give an appropriate error.

Reviewed By: wtaha

Differential Revision: D24461905

fbshipit-source-id: b88f60b40a6c9b7a367b33213ed0349e485fee27