Give compiler errors for stochastic comparisons (#450)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/450

We need to give an error when an attempt is made to compare anything with a stochastic node, because we have no ability at this time to compile it into a BMG graph. Here we turn statements of the form  `x = y op z` for op being any of `< > <= >= == !=` into graph nodes in the graph accumulator, and then detect in the problem fixer whether any such nodes remain and give an appropriate error.

Reviewed By: wtaha

Differential Revision: D24461905

fbshipit-source-id: c9e74ebb7dbfdf6dc6c3b4f0864075da776adb36