Start adding variance equality hypothesis test
Summary: This is the first of a series of diffs that will introduce a hypothesis test for the equality of variance. It is expected that the structure of this test will parallel that of the mean equality hypothesis test. However, with the experience of building that implementation, we aim to introduce this implementation with smaller diffs tha are easier to review.

Differential Revision: D24605829

fbshipit-source-id: 1fb31736242a879bc0f11067451db6108f62699f