Rewrite world as a context (#463)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/463

This diff is an attempt to see if we can move away from a global "world" and use context to limit its scope.

Usage:
```
w = World()
with w:
    # calls to random variables or the wrapper functions
    # are going to be associated with `w`
    foo()
```

Additionally, when a world is activated, the outer context is kept and will be restored when we leave the context, which enable us to do something like
```
with World() as w1:
    # do something to update w1

    with World() as w2:
       # do something else to update w2

    # w1 context is restored, we can continue updating it if needed
```

The context is only necessary for function that invokes the `random_variable` or `functional` decorator. To simplify the process, I also added a method `World.get()` that activate the context and invoke the random variable.

So instead of
```
StatisticalModel.set_world(world)
StatisticalModel.set_mode(Mode.INFERENCE)
value = get_wrapper(node.function)(*node.arguments)
```
we could simply do
```
value = world.get(node)
```

Reviewed By: nimar, yucenli

Differential Revision: D24916384

fbshipit-source-id: 944097e959ed51b8a4b1f82c675484d2ae441730