Fixing world.set_all_nodes_transform and set_all_nodes_proposer (#468)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/468

Seems like in our current implementation, when we set the transform and proposer for all nodes, we tried to preserve existing settings before setting the global default. However, in line 155 and 184, we effectively override what we just assigned in the few lines before. I'm fixing it in this diff by holding the value in a temporary dict so that they're not being override.

Differential Revision: D25149200

fbshipit-source-id: d95388d9223717da638b73b131250d9d2a04eed1