Introduce RVWrapper (#469)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/469

TL;DR: This diff introduce the `RVWrapper` *class*, which replaces [our current wrapper *function*](https://fburl.com/diffusion/19jdj3gi). It allows us to

- check the object that the random variable function belongs to (with `RVWrapper.model`)
- makes it easier for us to distinguish a decorated function from a regular Python function (with `isinstance(foo, RVWrapper)`

The former is especially important for serialization (WIP), because we want to avoid name collision in different models.

(Feel free to suggest another name for the wrapper class. I'm calling it `RVWrapper` just because we already use the term `wrapper` in a few places to denote the thing that decorate the  original function)

(Note on compositional inference: it seems like we were using the unbounded function to denote random variables that should be put into the same block (i.e., if `m = SomeModel()` and a user says `m.foo` should be a block, then this applies to all instances of `SomeModel`). I'm not quite sure whether this is intentional or not... I made some changes so that whatever user provides only applies to the specific instance, but it's very easy to go back if that's what we actually want.

-----

As for why this is needed to obtain a reference to the object:

Currently, we define `random_variable` decorator roughly as follows:
```
def random_variable(f):
    wraps(f)
    def wrapper(*args):
        # updating graph, etc.
    return wrapper
```
When using it inside of a class, the decorator is called at declaration time. e.g., in the following case
```
class SomeModel:
    random_variable
    def foo(self, x):
        return Normal(0, 1)
```
The function `wrapper` effectively replace `foo` as an attribute of class `SomeModel`. In fact, if we weren't [using `wraps(f)` to update the metadata](https://docs.python.org/3/library/functools.html), `SomeModel.foo.__name__` would be `"wrapper"` instead of `"foo"`.

Because all of these happen at declaration time, `SomeModel.foo` is not bound to any instance and it has type `"function"`. Then,

```
model = SomeModel()
model.foo  # <- here
```
On the line denoted as "here", the `__get__` method of our `wrapper` function is invoked, which *creates a new `"method"` that's bound to `model`* (i.e. the first argument that we usually called `self` is now a reference to `model`). Notice that this is does not modify `wrapper` itself but wraps around it ([see this doc for a better explanation](https://docs.python.org/3/howto/descriptor.html#functions-and-methods)). On the wrapper's perspective, it's still a regular `"function"`, it just that the first argument happens to be a reference to model.

Before the change, when we call `model.foo(5)`, `RVIdentifier` will store something like the follows:
- function: a reference of `SomeModel.foo`  (not bound to any instance
- arguments: a tuple `(model, 5)`

In general, we can't really distinguish whether the first arguments comes from a function binding or whether it's an explicit argument, this is why wrapper function doesn't work if we want to get a reference of the object.

After the change, `RVIdentifier` will store something like this
- function: a reference of the actual `model.foo` (already bound by `RVWrapper.__get__`
- arguments:  a tuple `(5,)`

and we could easily obtain a reference to `model` using the `__self__` attribute of the method.

Differential Revision: D25153759

fbshipit-source-id: ac6ab423f27048613587f84c8e5562d0afdae96a