Small refactoring in BM->BMG compiler
Summary:
I am embarking upon a stack of diffs with the intent of prototyping up a "JIT" compilation mechanism for our Bean Machine-to-Bean Machine Graph compiler.

Today the compiler operates on only a single, entire module and presupposes that this entire module will be transformed and evaluated in order to produce the graph. This is insufficiently granular and insufficiently powerful to enable what we really want: use of BMG as "just another inference engine".

The technique I am exploring to fix this is to make the compiler instead use a "just in time" mechanism. Just as the JavaVM or .NET CLR jitters compile each individual method when it is first encountered, so too can we transform *individual methods* into a form that accumulates graph nodes into a builder. As each method executes, it then transforms the methods it calls into the lifted form, and then caches the results.

In this diff I am beginning simply; I'm just refactoring the compilation method so that I can use it in a mechanism that compiles individual functions rather than entire modules.  That mechanism will be built up in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D24874698

fbshipit-source-id: 48f562795adf604d78fb1f46e113f714f98abbcf