Start building JIT compiler for Beanstalk
Summary:
As noted in the previous diff, I am prototyping up a "just in time" style compiler; here we have a helper method that I will need. This method takes a reference to a function, obtains its source code, transforms it into the lifted form, and then captures it inside an outer helper function which closes it over a graph builder.

In the next diff I will then compile this new method inside the module the original came from, and obtain a reference to the lifted function.

Reviewed By: wtaha

Differential Revision: D24904328

fbshipit-source-id: d702a84f14319bc9a1bdc1596e0f9bba49166d74