Install a patched version of pytest on CircleCI (#477)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/477

As titled. Currently pytest's `import-mode=importlib` option does not add the test file to `sys.module`, which can break some of the use cases. While [the community has came up with a fix](https://github.com/pytest-dev/pytest/pull/7870), it's not yet clear when the particular PR will be merged and added to the latest pytest. So in the meanwhile, we could download and use the patched version by directly from the PR so we are not blocked.

Differential Revision: D25480894

fbshipit-source-id: 761b882179d39b7919088a787b6c10e2e03abb77