Continue building parts for JIT compiler (#476)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/476

In the previous diff we showed that we could generate an AST for a lifted function that was closed over a graph builder.  In this diff we build on this work in a few ways:

* The function decorators put on random variable functions are now imported inside the scope of the helper function, thereby decreasing the amount of "pollution" of the outer namespace further.
* We now compile the generated source code and execute it. The global variable context for the execution is that of the *original* module, so if the lifted function makes use of a global, it should still work.
* I've moved the test cases to their own module.
* The test case is extended to show the code generation of both a normal function and a random variable
* The test case calls the functions and verifies that the supplied graph builder accumulates the graph.

The next steps will be to obtain the random variable id object returned during execution of a model, turn that into a lifted function and use it to accumulate the random variable into a graph builder.

Reviewed By: wtaha

Differential Revision: D24938503

fbshipit-source-id: cf665eec443f83d7ba4ef886fbd371b851ab365c