A Picklable RVIdentifier (#472)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/472

It turns out that it's actually possible to serialize RVIdentifier without computing the unique string-based keys and [enforcing class-based models](https://fb.quip.com/UFTPA8lnJiyL)... and **all it requires is changing the reference to `function` to a reference to `wrapper` in `RVIdentifier` (see `rv_identifier.py`)**. The rest of the diff is just some testing to ensure that pickle really works as expected.

----

Usage:

We should be able to pickle & unpickle `RVIdentifier` (and `MonteCarloSamples`) in general like follows:

```
bm.random_variable
def foo():
    return dist.Normal(0., 1.)

mc = SomeMCMCAlgorithm()
samples = mc.infer(queries=[foo()], observations={}, **other_args)

with open('some_file.pkl', 'wb') as f:
    pickle.dump(samples, f)
```
and in another session
```
# we still need to have the function in scope when unpickling
bm.random_variable
def foo():
    return dist.Normal(0., 1.)

with open('some_file.pkl', 'rb') as f:
    samples = pickle.load(f)

samples[foo()]  # <- this should still work
```

For random variables defined inside of classes, we'll need to pickle the model together with the samples to keep the reference. e.g.,
```
class SomeModel:
    bm.random_variable
    def foo(self):
        return dist.Normal(0., 1.)

model = SomeModel()
mc = SomeMCMCAlgorithm()
samples = mc.infer(queries=[model.foo()], observations={}, **other_args)

with open('some_file.pkl', 'wb') as f:
    # a tuple containing both the model and the samples are pickled
    # other formats (e.g., dict) could also be used here
    pickle.dump((model, samples), f)
```
(We can also use a Pickler instance if we want to keep the reference but store model and samples in separate files: https://stackoverflow.com/questions/6376081/pickle-linked-objects)

Then, in another session:
```
# similarly, the definition of the class needs to be in scope when unpickling
class SomeModel:
    bm.random_variable
    def foo(self):
        return dist.Normal(0., 1.)

with open('some_file.pkl', 'rb') as f:
    model, samples = pickle.load(f)

samples[model.foo()]  # <- this should still work
```
----

Some explanation for those that are interested:

Previously, we were not able to pickle the RVIdentifier, because by default, Python serialize class instances by serializing its `__dict__`, which contains a reference to the original `function` (the one being decorated). However, Python serialize functions by its module and name. The issue is, when Python look up the inner function in this way, what it ended up finding is the wrapper function that shares the same name and in the same module as the inner one, e.g.,
```
PicklingError: Can't pickle <function foo at 0x7f2deb636170>: it's not the same object as __main__.foo
```

If we store a reference to wrapper instead, then this will not be an issue, because the reference to wrapper is recoverable given its name and module. Pickle does not store the actual body of the function, so when unpickling, we should still have the definition of the class/function in scope. The unit tests included in this diff contain some examples to showcase how we can use pickle in different scenarios.

Some final thoughts: string-based identifier might still be helpful, but they put more restrictions on the source of RVIdentifier. So instead of using it in serialization, it might be better to make it an optional feature...

Reviewed By: nimar

Differential Revision: D25421261

fbshipit-source-id: 1c78c1196c6acd999cf36715e439dd28ea7c73b5