Continue implementing JIT compiler for beanstalk (#479)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/479

I am continuing my incremental progress towards a JIT compiler in Beanstalk. In this diff, I extend the work in the previous diff by making a helper method on a graph builder which can turn an RVID into a sample node. It does so by:

* obtaining the function reference
* recompiling its source code into a helper method closed over a graph builder
* closing it over *this* graph builder
* invoking the lifted, closed-over function with the arguments captured in the RVID.

For simple cases where the RV just returns a distribution, this now works.  More complex RVs that involve calls to other RVs does not yet work.

In upcoming diffs:

* We'll build a cache to map RVID -> sample node
* We'll build a cache to map unlifted function -> lifted helper function

That will ensure that we do not compile the same functions over and over again when we encounter them, decreasing the amount of work the runtime must do.

Once that's done:

* We'll fix up function call handling in lifted code so that it, in turn, detects when it is calling a function that needs to be lifted.
* We'll add an inference method that takes RVIDs for queries and observations, jit-compiles them, and runs inference

Reviewed By: wtaha

Differential Revision: D24961042

fbshipit-source-id: 86605e67d49137df8f0e30af8eff06d18ececfa1