Build a cache to make rv->node computations idempotent
Summary:
For both performance and correctness we would like to ensure that as we jit-compile a model:

* an rv function should be compiled to its lifted form only once; this operation is work and we can avoid doing it twice.
* a lifted rv function always produces the same node for a given input, so we can cache that as well.

(The sharp-eyed reviewer may have noted that we already ensure the latter by causing the lifted rv function to be decorated with a "memoize" annotation, so it seems like we have some redundancy here. I will likely remove the memoize annotation in a later diff.)

The easiest way to achieve idempotency is with a cache; we actually build two caches. Since an RVIdentifier identifies both a function and its arguments, we can cache both the mapping from one RVID to a graph node, and from the unlifted function to the lifted function.  That way, when we see a new RVID that has the same function as a previous one, we can re-use the lifted function.

Reviewed By: wtaha

Differential Revision: D25027485

fbshipit-source-id: 5b9403ccced06a0b966b8b62d64d52da7e96e468