Build a cache to make rv->node computations idempotent (#482)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/482

An RVID represents a call to a random variable (or functional) function *with a particular argument list*. Our queries and observations sets use RVIDs to identify which random quantities we are querying or observing. Therefore, one of the basic tasks we need to perform in the compiler is translating an RVID into a node in a graph.

Doing so is potentially a lot of work: we have to first translate the body of the random variable function into its lifted form, then execute that code to populate the graph builder, and then obtain the graph node that is returned when it is invoked with the argument list of the RVID.

Therefore, for both performance and correctness we would like to ensure that as we jit-compile a model:

* We cache the code that is generated for an rv function, regardless of the arguments. We thereby ensure that we only do the compilation to the lifted form once.

* We cache the node that is returned when a lifted function is called with a particular arguments list.

(The sharp-eyed reviewer may have noted that we already ensure the latter by causing the lifted rv function to be decorated with a "memoize" annotation, so it seems like we have some redundancy here. I will likely remove the memoize annotation in a later diff.)

The easiest way to achieve idempotency is with a cache; we actually build two caches. Since an RVIdentifier identifies both a function and its arguments, we can make a map that caches the unlifted function -> lifted function work, and a map from RVID -> graph node.

That way, when we encounter an RVID for the second time, we already have the graph node computed from the first time. And if we encounter an RVID with different arguments but the same unlifted function as a previous RVID, we can at least skip the compilation step.

Reviewed By: wtaha

Differential Revision: D25027485

fbshipit-source-id: 846f6f18f214fc0ac314666294b17966a9c404ef