Mark random variable functions
Summary:
For my JIT compiler I'm going to need to know when a function object is a random variable / functional, *before* I execute it. The easiest way to do that is to simply have the wrapper function generated by the decorator mark itself as being a random variable / functional; that way I can just check whether there is such an attribute.

I'll be making use of this in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D25034512

fbshipit-source-id: b380f6d116ded1faa26f6c28ca6ad5ffc7436780