Implement simplest automatic jit-compilation scenario
Summary:
My work on a jit compiler continues. Here we have the simplest scenario: we have already compiled one random variable function, and it is calling a second, uncompiled random variable function with no arguments.

In this situation, we actually do the call on the uncompiled function to get back an RVIdentifier. We then use our existing jit-compiling mechanism to turn that into a node, cache it, and return the result.

This will be the first of many scenarios we need to handle.  We need to do the cross product of:

* callee: rv, functional, normal function with source code, function without source code
* arguments: all normal arguments, some graph nodes as arguments

That gives us eight scenarios to implement and test.

Reviewed By: wtaha

Differential Revision: D25042290

fbshipit-source-id: a661ad59b69a836268ddb3fdfbd7569d934eaddd