Handle jit compilation of "functionals" with ordinary arguments
Summary:
I'm continuing down the list of scenarios listed in my previous diff for jit-compiled function calls.

In this diff, an easy one. A call to a "functional" function with normal arguments (that is, not themselves graph nodes) is handled exactly like a "random variable" call with normal arguments.  In both cases the function will return an RVIdentifier that captures both the function and the arguments; we can use our existing mechanism for compiling, executing and caching the resulting graph node.

Note that we do not yet handle the case where a "functional" returns a non-graph node. We'll deal with that in a later diff.

Reviewed By: wtaha

Differential Revision: D25077894

fbshipit-source-id: b978f4cc94d8b1228a6d8f8c3a9e4a0dffa517ae