Remove automatic addition of query nodes to functionals in compiler
Summary:
When we translate a model from BeanMachine to BMGraph, we must insert a "sample" graph node onto every distinct call to a `random_variable` method.

When I was first implementing the compiler I did not understand how "query" nodes were properly to be inserted, and made the incorrect assumption that we must insert a query node on every distinct call to a `functional` method, similar to how we insert a sample on every `random_variable`.

This is incorrect. The `functional` annotation marks computations that *can* be queried, not computations that *are* queried. The only query nodes added to the graph builder ought to be those pointing to nodes that were actually queried during inference.

I have therefore removed the code that automatically inserts query nodes on every call to a `functional`.

In an upcoming diff we will show how the new JIT compilation mechanism inserts query nodes into the graph correctly.

Reviewed By: wtaha

Differential Revision: D25258391

fbshipit-source-id: df4af4faf4fea2e08d6c7ef70686eab7947d9c62