Test case to demonstrate rewriting bug
Summary:
I noticed a bug in the rewriting logic; for some reason, nested function calls used as named arguments inside a function call are not being rewritten correctly.

The test passes but it should not; once the bug is fixed, we can either delete the test or change it to pass.

Reviewed By: wtaha

Differential Revision: D25449184

fbshipit-source-id: eaa36bc50dd5b5305fcea870ac5dddbae507d3f5