Fix bugs in phi detection
Summary:
We need to be able to tell if the phi function is called with a graph node as its argument; for our purposes the phi function is any function of the form `Normal(0, 1).cdf`.

I wrote a little method to detect when this function is the one being called but I made two very silly mistakes. The first is: I check the function to see if it is a member of `Normal` by checking `__self__`, but not all function objects have a `__self__` attribute, so this can crash. The second is: I never checked the name of the function! All member functions of `Normal(0, 1)` were treated as phi. Whoops.

Reviewed By: wtaha

Differential Revision: D25279291

fbshipit-source-id: 5842d3b90f69daed99d390f3374ef7da8f1f4b66