Test case to demonstrate rewriting bug (#491)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/491

I noticed a bug in the rewriting logic; for some reason, nested function calls used as named arguments inside a function call are not being rewritten correctly.

The test passes but it should not; once the bug is fixed, we can either delete the test or change it to pass.

Reviewed By: wtaha

Differential Revision: D25449184

fbshipit-source-id: 4e7f486ff0e296ab0568c8aca1fbd432dd0d4c16