Clarified prior comments to make it clear that they document issues that have been fixed
Summary: This is a clean up diff on a stack that starts with D25449184 (https://github.com/facebookincubator/beanmachine/commit/9c89de82aaf113f5978874c6f3fee53d9adfad17) where an issue was identified with expressions of the form "x = dict(...)" that are generated by rewrites that handle keyword arguments. This diff simply clariied comments made earlier int he stack to make sure that it is clear that the issues raised in those comments are not already fixed.

Reviewed By: ericlippert

Differential Revision: D25616696

fbshipit-source-id: 4dd1b27ac562fb47e3495627389e1cdea46e99d7