Fix to problem with unary applications of dict
Summary: This diff fixed the problem identified in D25449184 (https://github.com/facebookincubator/beanmachine/commit/9c89de82aaf113f5978874c6f3fee53d9adfad17), namely, that the arguments to a dict(...) are not simplified in the case when ... is n=expression. This diff fixes that particular problem, thereby implementing the first of three rewrites that are identified in the bug fix overview diff D25606994.

Reviewed By: ericlippert

Differential Revision: D25614876

fbshipit-source-id: 7c61790026c2e4e8cd3bb18fccc227873c95c10f