Some more tests to diagnose this issue
Summary:
This diff is a followup to a bug report diff D25449184 (https://github.com/facebookincubator/beanmachine/commit/9c89de82aaf113f5978874c6f3fee53d9adfad17). Here I am just documenting some additional observations relating to this issue. It seems that what is missing is rewrites that that compensate for the fact that "dict(..)" is currently an "allowed_function", which means that it is not really processed any further. At this point it seems that we will need at least two more rewrites:
  x = dict(n=complicated) -> y = complicated; x = dict(n=y)
and maybe more importantly,
  x = dict(**c,**e) -> y = c; x = dict(**y,**e)
and
   x = dict(**n,**c) -> y = c; x = dict(**n,**y)
A nice thing about those two sets of rewrites is that the first is similar to unaryop rewrites and the second set is similar to binary op rewrites.

Reviewed By: ericlippert

Differential Revision: D25606994

fbshipit-source-id: 38e3d17d0b0a0d126590aab79ba4c67a855e6283