Start adding variance equality hypothesis test (#437)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/437

This is the first of a series of diffs that will introduce a hypothesis test for the equality of variance. It is expected that the structure of this test will parallel that of the mean equality hypothesis test. However, with the experience of building that implementation, we aim to introduce this implementation with smaller diffs tha are easier to review.

In this particular diff we simply bring in the CDF for the chi-squared distribution and check it on some typical probabilities and a df of 100 against values in a standard table.

Reviewed By: SepehrAkhavan

Differential Revision: D24605829

fbshipit-source-id: f5e3b3891c5c760ec87785cd320438c657cd0fa7