Minor updates to comments and order of code for clarity
Summary: The file hypothesis_testing.py was previously just for testing means, so, the comment is changed to prepare for the addition of the second test. Also, the hypothesis testing function is placed before the function for constructing the confidence interval, as the first more directly presented the idea of the tests.

Differential Revision: D24609581

fbshipit-source-id: 03f785570081ab3512dcb5075cf211e3ede0cd82