Fix bugs in phi detection (#492)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/492

We need to be able to tell if the phi function is called with a graph node as its argument; for our purposes the phi function is any function of the form `Normal(0, 1).cdf`.

I wrote a little method to detect when this function is the one being called but I made two very silly mistakes. The first is: I check the function to see if it is a member of `Normal` by checking `__self__`, but not all function objects have a `__self__` attribute, so this can crash. The second is: I never checked the name of the function! All member functions of `Normal(0, 1)` were treated as phi. Whoops.

Reviewed By: wtaha

Differential Revision: D25279291

fbshipit-source-id: f8e6a37a2fe82dadadeaf7b2cb018be85c6831d9