Increase delta for NUT distant normal normal conjugate tests
Summary:
As titled. This is just a temporary fix so that our on-call won't receive a new task regarding test error every night (especially not during the holidays {emoji:1f604}).

We should still investigate what happen behind the scene (yucenli). Though it's also possible that our current delta tests just happen to be bad, in which case Walid's hypothesis testing work could be sufficient in fixing the issue permanently

Differential Revision: D25652964

fbshipit-source-id: b80a7ed8b410f68576ff0e94d04454cabc145644